+++ packages/node-core/src/configure/NodeConfig.ts
@@ -31,6 +31,8 @@ export interface IConfig {
   readonly proofOfIndex: boolean;
   readonly mmrPath?: string;
   readonly ipfs?: string;
+  readonly dictionaryTimeout: number;
+  readonly workers?: number;
 }
 
 export type MinConfig = Partial<Omit<IConfig, 'subquery'>> & Pick<IConfig, 'subquery'>;
@@ -45,6 +47,7 @@ const DEFAULT_CONFIG = {
   indexCountLimit: 10,
   timestampField: true,
   proofOfIndex: false,
+  dictionaryTimeout: 30,
 };
 
 export class NodeConfig implements IConfig {
@@ -103,6 +106,7 @@ export class NodeConfig implements IConfig {
   get debug(): boolean {
     return this._config.debug;
   }
+
   get preferRange(): boolean {
     return this._config.preferRange;
   }
@@ -131,6 +135,10 @@ export class NodeConfig implements IConfig {
     return this._config.proofOfIndex;
   }
 
+  get dictionaryTimeout(): number {
+    return this._config.dictionaryTimeout;
+  }
+
   get mmrPath(): string {
     return this._config.mmrPath ?? `.mmr/${this.subqueryName}.mmr`;
   }
@@ -140,6 +148,10 @@ export class NodeConfig implements IConfig {
 
   get dbSchema(): string {
     return this._config.dbSchema ?? this.subqueryName;
+  }
+
+  get workers(): number {
+    return this._config.workers;
   }
 
   merge(config: Partial<IConfig>): this {
