+++ packages/node-core/src/meta/ready.service.ts
@@ -1,9 +1,9 @@
 // Copyright 2020-2022 OnFinality Limited authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import {Injectable} from '@nestjs/common';
-import {OnEvent} from '@nestjs/event-emitter';
-import {EventPayload, IndexerEvent} from '../events';
+import { Injectable } from '@nestjs/common';
+import { OnEvent } from '@nestjs/event-emitter';
+import { EventPayload, IndexerEvent } from '../events';
 
 @Injectable()
 export class ReadyService {
@@ -14,11 +14,11 @@ export class ReadyService {
   }
 
   @OnEvent(IndexerEvent.Ready)
-  handleReady({value}: EventPayload<boolean>): void {
+  handleReady({ value }: EventPayload<boolean>): void {
     this._ready = value;
   }
 
-  get ready() {
+  get ready(): boolean {
     return this._ready;
   }
 }
