+++ packages/node-core/src/utils/graphql.ts
@@ -1,11 +1,22 @@
 // Copyright 2020-2022 OnFinality Limited authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import {isHex, hexToU8a, u8aToBuffer, u8aToHex, bufferToU8a, isBuffer, isNull} from '@polkadot/util';
-import {getTypeByScalarName, GraphQLModelsType} from '@subql/utils';
-import {ModelAttributes, ModelAttributeColumnOptions} from 'sequelize';
+import {
+  isHex,
+  hexToU8a,
+  u8aToBuffer,
+  u8aToHex,
+  bufferToU8a,
+  isBuffer,
+  isNull,
+} from '@polkadot/util';
+import { getTypeByScalarName, GraphQLModelsType } from '@subql/utils';
+import { ModelAttributes, ModelAttributeColumnOptions } from 'sequelize';
 
-export function modelsTypeToModelAttributes(modelType: GraphQLModelsType, enums: Map<string, string>): ModelAttributes {
+export function modelsTypeToModelAttributes(
+  modelType: GraphQLModelsType,
+  enums: Map<string, string>,
+): ModelAttributes {
   const fields = modelType.fields;
   return Object.values(fields).reduce((acc, field) => {
     const allowNull = field.nullable;
@@ -22,10 +33,20 @@ export function modelsTypeToModelAttributes(modelType: GraphQLModelsType, enums:
     if (field.type === 'BigInt') {
       columnOption.get = function () {
         const dataValue = this.getDataValue(field.name);
+        if (field.isArray) {
+          return dataValue ? dataValue.map((v) => BigInt(v)) : null;
+        }
         return dataValue ? BigInt(dataValue) : null;
       };
       columnOption.set = function (val: unknown) {
-        this.setDataValue(field.name, val?.toString());
+        if (field.isArray) {
+          this.setDataValue(
+            field.name,
+            (val as unknown[])?.map((v) => v.toString()),
+          );
+        } else {
+          this.setDataValue(field.name, val?.toString());
+        }
       };
     }
     if (field.type === 'Bytes') {
@@ -35,7 +56,9 @@ export function modelsTypeToModelAttributes(modelType: GraphQLModelsType, enums:
           return null;
         }
         if (!isBuffer(dataValue)) {
-          throw new Error(`Bytes: store.get() returned type is not buffer type`);
+          throw new Error(
+            `Bytes: store.get() returned type is not buffer type`,
+          );
         }
         return u8aToHex(bufferToU8a(dataValue));
       };
@@ -46,7 +69,9 @@ export function modelsTypeToModelAttributes(modelType: GraphQLModelsType, enums:
           const setValue = u8aToBuffer(hexToU8a(val));
           this.setDataValue(field.name, setValue);
         } else {
-          throw new Error(`input for Bytes type is only support unprefixed hex`);
+          throw new Error(
+            `input for Bytes type is only support unprefixed hex`,
+          );
         }
       };
     }
