+++ packages/node-core/src/utils/promise.ts
@@ -1,6 +1,7 @@
 // Copyright 2020-2022 OnFinality Limited authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
+//deprecate method, use delay from @subql/common
 export async function delay(sec: number): Promise<void> {
   return new Promise((resolve) => {
     setTimeout(resolve, sec * 1000);
