+++ packages/node/package.json
@@ -27,6 +27,7 @@
     "@polkadot/api": "9.1.1",
     "@subql/common": "workspace:*",
     "@subql/common-substrate": "workspace:*",
+    "@subql/node-core": "workspace:*",
     "@subql/types": "workspace:*",
     "@subql/utils": "workspace:*",
     "@subql/x-merkle-mountain-range": "2.0.0-0.1.2",
