+++ packages/node/src/app.module.ts
@@ -5,7 +5,7 @@ import { Module } from '@nestjs/common';
 import { EventEmitterModule } from '@nestjs/event-emitter';
 import { ScheduleModule } from '@nestjs/schedule';
 import { ConfigureModule } from './configure/configure.module';
-import { DbModule } from './db/db.module';
+import { DbModule } from '@subql/node-core/db';
 import { FetchModule } from './indexer/fetch.module';
 import { MetaModule } from './meta/meta.module';
