--- packages/node/src/configure/NodeConfig.spec.ts
@@ -1,38 +1,0 @@
-// Copyright 2020-2022 OnFinality Limited authors & contributors
-// SPDX-License-Identifier: Apache-2.0
-
-import * as path from 'path';
-import { NodeConfig } from './NodeConfig';
-
-describe('NodeConfig', () => {
-  it('supports read from yaml', () => {
-    const config = NodeConfig.fromFile(
-      path.join(__dirname, '../../test/config.yml'),
-    );
-    expect(config).toBeInstanceOf(NodeConfig);
-    expect(config).toMatchObject({
-      subquery: '../../../../subql-example/extrinsics',
-      subqueryName: 'extrinsics',
-    });
-  });
-
-  it('support read from json', () => {
-    const config = NodeConfig.fromFile(
-      path.join(__dirname, '../../test/config.json'),
-    );
-    expect(config).toBeInstanceOf(NodeConfig);
-    expect(config).toMatchObject({
-      subquery: '../../../../subql-example/extrinsics',
-      subqueryName: 'extrinsics',
-    });
-  });
-
-  it('throw error for unknown configs', () => {
-    expect(() =>
-      NodeConfig.fromFile(path.join(__dirname, '../../test/config.toml')),
-    ).toThrow();
-    expect(() =>
-      NodeConfig.fromFile(path.join(__dirname, '../../test/con.toml')),
-    ).toThrow(/Load config from file/);
-  });
-});
