--- packages/node/src/entities/index.ts
@@ -1,4 +1,0 @@
-// Copyright 2020-2022 OnFinality Limited authors & contributors
-// SPDX-License-Identifier: Apache-2.0
-
-export * from './Subquery.entity';
