+++ packages/node/src/indexer/fetch.service.spec.ts
@@ -11,7 +11,7 @@ import {
   SubstrateHandlerKind,
 } from '@subql/common-substrate';
 import { GraphQLSchema } from 'graphql';
-import { NodeConfig } from '../configure/NodeConfig';
+import { NodeConfig } from '@subql/node-core/configure';
 import { SubqueryProject } from '../configure/SubqueryProject';
 import { calcInterval, fetchBlocksBatches } from '../utils/substrate';
 import { ApiService } from './api.service';
