+++ packages/node/src/indexer/fetch.service.test.ts
@@ -12,7 +12,7 @@ import {
   SubstrateHandlerKind,
 } from '@subql/common-substrate';
 import { GraphQLSchema } from 'graphql';
-import { NodeConfig } from '../configure/NodeConfig';
+import { NodeConfig } from '@subql/node-core/configure';
 import { SubqueryProject } from '../configure/SubqueryProject';
 import { delay } from '../utils/promise';
 import * as SubstrateUtil from '../utils/substrate';
