+++ packages/node/src/indexer/fetch.service.ts
@@ -21,24 +21,23 @@ import {
   SubstrateRuntimeHandlerFilter,
   SubstrateBlockFilter,
 } from '@subql/common-substrate';
+import { getYargsOption } from '@subql/node-core';
+import { NodeConfig } from '@subql/node-core/configure';
+import { BlockedQueue } from '@subql/node-core/indexer';
+import { delay } from '@subql/node-core/utils';
 import {
   DictionaryQueryEntry,
   SubstrateBlock,
   SubstrateCustomHandler,
 } from '@subql/types';
-
 import { MetaData } from '@subql/utils';
-import { range, sortBy, uniqBy } from 'lodash';
-import { NodeConfig } from '../configure/NodeConfig';
+import { isUndefined, range, sortBy, uniqBy } from 'lodash';
 import { SubqlProjectDs, SubqueryProject } from '../configure/SubqueryProject';
 import { checkMemoryUsage } from '../utils/batch-size';
 import { getLogger } from '../utils/logger';
 import { profiler } from '../utils/profiler';
 import { isBaseHandler, isCustomHandler } from '../utils/project';
-import { delay } from '../utils/promise';
 import * as SubstrateUtil from '../utils/substrate';
-import { calcInterval } from '../utils/substrate';
-import { getYargsOption } from '../yargs';
 import { ApiService } from './api.service';
 import { DictionaryService, SpecVersion } from './dictionary.service';
 import { DsProcessorService } from './ds-processor.service';
