+++ packages/node/src/indexer/poi.service.ts
@@ -4,10 +4,10 @@
 import { isMainThread } from 'node:worker_threads';
 import { Injectable, OnApplicationShutdown } from '@nestjs/common';
 import { hexToU8a } from '@polkadot/util';
+import { NodeConfig } from '@subql/node-core/configure';
 import { Sequelize } from 'sequelize';
-import { NodeConfig } from '../configure/NodeConfig';
 import { SubqueryProject } from '../configure/SubqueryProject';
-import { PoiFactory, PoiRepo } from './entities/Poi.entity';
+import { PoiFactory, PoiRepo } from '@subql/node-core/indexer/entities';
 
 const DEFAULT_PARENT_HASH = hexToU8a('0x00');
