+++ packages/node/src/indexer/project.service.ts
@@ -8,15 +8,15 @@ import { Inject, Injectable } from '@nestjs/common';
 import { EventEmitter2, OnEvent } from '@nestjs/event-emitter';
 import { getAllEntitiesRelations } from '@subql/utils';
 import { QueryTypes, Sequelize } from 'sequelize';
-import { NodeConfig } from '../configure/NodeConfig';
+import { NodeConfig } from '@subql/node-core/configure';
 import { SubqlProjectDs, SubqueryProject } from '../configure/SubqueryProject';
-import { SubqueryRepo } from '../entities';
+import { SubqueryRepo } from '@subql/node-core/entities';
 import { getLogger } from '../utils/logger';
-import { getYargsOption } from '../yargs';
+import { getYargsOption } from '@subql/node-core';
 import { ApiService } from './api.service';
 import { DsProcessorService } from './ds-processor.service';
 import { DynamicDsService } from './dynamic-ds.service';
-import { MetadataFactory, MetadataRepo } from './entities/Metadata.entity';
+import { MetadataFactory, MetadataRepo } from '@subql/node-core/indexer/entities';
 import { IndexerEvent } from './events';
 import { MmrService } from './mmr.service';
 import { PoiService } from './poi.service';
