+++ packages/node/src/indexer/sandbox.service.ts
@@ -7,16 +7,16 @@ import {
   isDatasourceV0_2_0,
   SubstrateDataSource,
 } from '@subql/common-substrate';
+import { NodeConfig } from '@subql/node-core/configure';
 import { Store } from '@subql/types';
 import { levelFilter } from '@subql/utils';
 import { merge } from 'lodash';
 import { NodeVM, NodeVMOptions, VMScript } from 'vm2';
-import { NodeConfig } from '../configure/NodeConfig';
 import { SubqlProjectDs, SubqueryProject } from '../configure/SubqueryProject';
 import { getLogger } from '../utils/logger';
 import { getProjectEntry } from '../utils/project';
-import { timeout } from '../utils/promise';
-import { getYargsOption } from '../yargs';
+import { timeout } from '@subql/node-core/utils';
+import { getYargsOption } from '@subql/node-core';
 import { ApiService } from './api.service';
 import { StoreService } from './store.service';
 import { ApiAt } from './types';
