+++ packages/node/src/main.ts
@@ -3,12 +3,12 @@
 
 import { NestFactory } from '@nestjs/core';
 import { findAvailablePort } from '@subql/common';
+import { getYargsOption } from '@subql/node-core';
 import { AppModule } from './app.module';
 import { ApiService } from './indexer/api.service';
 import { FetchService } from './indexer/fetch.service';
 import { ProjectService } from './indexer/project.service';
 import { getLogger, NestLogger } from './utils/logger';
-import { getYargsOption } from './yargs';
 
 const DEFAULT_PORT = 3000;
 const logger = getLogger('subql-node');
